rofs.rb: add description.

+++ utils/rofs.rb
@@ -21,15 +21,27 @@
 # THE SOFTWARE.
 ################################################################################
 
+# rofs.rb:
+#   Prepare /rofs
+#   - Prepare ESSENTIAL_DIRS
+#   - Mirror MIRROR_DIRS and mark read-only to /rofs
+#   - Prepare a `/etc/passwd` file
+#   - Mount a tmpfs at `/tmp`
+#   - Prepare `/dev/#{*DEV_NODES.keys}`
+
+ESSENTIAL_DIRS = ['/usr', '/bin', '/opt', '/lib', '/lib64', '/etc', '/dev', '/tmp', '/proc']
+MIRROR_DIRS    = ['/usr', '/bin', '/opt', '/lib', '/lib64']
+DEV_NODES      = {null: 3, zero: 5, random: 8, urandom: 9, full: 7}
+
+ROFS_DEST      = ENV['ROFS_DEST'] || '/rofs'
+TMPFS_SIZE     = ENV['ROFS_TMPSIZE'] || '49152k'
+
 if Process.uid != 0
   puts 'root required'
   exit 1
 end
 
 require 'fileutils'
-
-ROFS_DEST  = ENV['ROFS_DEST'] || '/rofs'
-TMPFS_SIZE = ENV['ROFS_TMPSIZE'] || '49152k'
 
 def assert_execute(*cmd)
   print ['>', *cmd, '... '].join(' ')
@@ -74,7 +86,7 @@ end
 
 FileUtils.mkdir_p ROFS_DEST
 
-['/usr', '/bin', '/opt', '/lib', '/lib64', '/etc', '/dev', '/tmp', '/proc'].each do |x|
+ESSENTIAL_DIRS.each do |x|
   dir = File.join(ROFS_DEST, x)
   next unless Dir.exists?(x) && (! Dir.exists?(dir))
   FileUtils.mkdir_p dir
@@ -89,27 +101,33 @@ end
 end
 
 # bind rofs
-['/usr', '/bin', '/opt', '/lib', '/lib64'].each do |src|
+MIRROR_DIRS.each do |src|
   dest = File.join(ROFS_DEST, src)
   next unless Dir.exists?(src) && Dir.exists?(dest)
   mount src, dest, '--bind', 'defaults', 'ro,noatime,nosuid,nodev'
 end
 
 # devfs
-dev_path = File.join(ROFS_DEST, 'dev')
-File.unlink *Dir[File.join(dev_path, '*')]
+if ESSENTIAL_DIRS.include?('/dev')
+  dev_path = File.join(ROFS_DEST, 'dev')
+  File.unlink *Dir[File.join(dev_path, '*')]
+end
 
-{null: 3, zero: 5, random: 8, urandom: 9, full: 7}.each do |dev, id|
+DEV_NODES.each do |dev, id|
   assert_execute 'mknod', '-m', '666', File.join(dev_path, dev.to_s), 'c', '1', id.to_s
 end
 
 # passwd
-File.open(File.join(ROFS_DEST, 'etc/passwd'), 'w') do |f|
-  f.puts "guest:x:2000:200:bin:/bin:/bin/false"
+if ESSENTIAL_DIRS.include?('/etc')
+  File.open(File.join(ROFS_DEST, 'etc/passwd'), 'w') do |f|
+    f.puts "guest:x:2000:200:bin:/bin:/bin/false"
+  end
 end
 
 # tmpfs
-mount 'none', File.join(ROFS_DEST, 'tmp'), ['-t', 'tmpfs'], "size=#{TMPFS_SIZE},relatime,nosuid"
+if ESSENTIAL_DIRS.include?('/tmp')
+  mount 'none', File.join(ROFS_DEST, 'tmp'), ['-t', 'tmpfs'], "size=#{TMPFS_SIZE},relatime,nosuid"
+end
 
 # done
 puts "ROFS: #{ROFS_DEST} is ready."
